% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfdr.R
\name{filter_data}
\alias{filter_data}
\title{Filter data according to thresholds.}
\usage{
filter_data(
  Object,
  peptide_centric = TRUE,
  LFDR_threshold = c("strong", "substantial"),
  min_nb_timepoints = 0,
  condition_names = colnames(Object@design)[colnames(Object@design) !=
    Object@time_unit],
  min_labeling_ratio = 0,
  min_nb_sample = NULL,
  score_threshold = NULL,
  higher_score_better = T,
  trace = T
)
}
\arguments{
\item{Object}{An object of class MetaProfiler.}

\item{peptide_centric}{Logical value specifying if filtering is done at the peptide or protein level. Protein level filtering is not yet fully implemented.}

\item{LFDR_threshold}{Either a number or a character value "strong" or "substantial". If "strong" is chosen, then the threshold is 0.0909 (odds for false discoveries is 1 out of 10). If "substantial" is chosen, then the LFDR threshold is 0.2 (odds for false discovery is 1 out of 3).}

\item{min_nb_timepoints}{Numeric value specifying the minimum number of timepoints the peptide or protein must be present in.}

\item{condition_names}{A character vector specifying the conditions used in the experiment. Names must match the columns in Object@design. Defaults to all columns in Object@design, except for the column containing the timepoints.  Names must match the columns in Object@design.}

\item{min_labeling_ratio}{(Only used when \code{estimate_prior_probability} is set to "mle") A vector with two strings values. The first specifies the name of the distribution of RIA for true discoveries and the second specifies the distribution of LR for true discoveries. Can be either "weibull", "beta", "norm", "lnorm", "exp", "gamma", "invtrgamma", "trgamma", "lgamma", "pareto", "pareto1", "invweibull", "llogis", "invgamma", "unif", "cauchy", or "logis".}

\item{min_nb_sample}{Numeric value specifying the minimum number of conditions the peptide or protein must be present in.}

\item{score_threshold}{The score threshold.}

\item{higher_score_better}{Logical value denoting if higher score is better.}

\item{trace}{If \code{TRUE}, then progress is printed. Defaults to \code{TRUE}.}
}
\value{
Returns an object of class MetaProfiler.

Returns an object of class MetaProfiler.

Returns an object of class MetaProfiler.
}
\description{
Filter data according to thresholds.
}
