% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfdr.R
\name{lfdr}
\alias{lfdr}
\title{Computes the LFDR of heavy features quantified from Protein-SIP experiments.}
\usage{
lfdr(
  Object,
  condition_names = NULL,
  estimate_prior_probability = c("mle", "analytically"),
  distr = rep("weibull", 2),
  trace = T
)
}
\arguments{
\item{Object}{An object of class MetaProfiler.}

\item{condition_names}{A character vector specifying the conditions used in the experiment. Names must match the columns in Object@design. If \code{NULL}, then all the conditions are combined to compute LFDR.}

\item{estimate_prior_probability}{Either "mle" or "analytically". If "mle" is chosen, prior probability for false discovery, p0, is estimated from \eqn{f(x) = f0(x)*p0 + f1(x)*(p0-1)} using maximum likelihood estimation, where \eqn{f0(x)}, \eqn{f1(x)}, and \eqn{f(x)} is the density of the data for false, true, and mixed discoveries, respectively. If "analytically" is chosen, prior probability is simply the number of features at time 0 of the corresponding \code{by} group divided by the number of features in the current group.}

\item{distr}{(Only used when \code{estimate_prior_probability} is set to "mle") A vector with two strings values. The first specifies the name of the distribution of RIA for true discoveries and the second specifies the distribution of LR for true discoveries. Can be either "weibull", "beta", "norm", "lnorm", "exp", "gamma", "invtrgamma", "trgamma", "lgamma", "pareto", "pareto1", "invweibull", "llogis", "invgamma", "unif", "cauchy", or "logis".}

\item{trace}{If \code{TRUE}, then progress is printed. Defaults to \code{TRUE}.}
}
\value{
Returns an object of class MetaProfiler.
}
\description{
Computes the LFDR of heavy features quantified from Protein-SIP experiments.
}
