\name{MetaProfiler-package}
\alias{MetaProfiler-package}
\alias{MetaProfiler}
\docType{package}
\title{\packageTitle{MetaProfiler}}
\description{\packageDescription{MetaProfiler}}
\section{Package Content}{\packageIndices{MetaProfiler}}
\author{\packageAuthor{MetaProfiler}}
\section{Maintainer}{\packageMaintainer{MetaProfiler}}
\keyword{package}
