% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfdr.R
\name{start_arg_default}
\alias{start_arg_default}
\title{Computes the default argument of various distributions. Adapted from fitdistrplus.}
\usage{
start_arg_default(x, distr)
}
\arguments{
\item{x}{A numeric vector}

\item{distr}{The name of the distribution. Can be either "weibull", "beta", "norm", "lnorm", "exp", "gamma", "invtrgamma", "trgamma", "lgamma", "pareto", "pareto1", "invweibull", "llogis", "invgamma", "unif", "cauchy", or "logis".}
}
\value{
A list of starting values for \code{distr}
}
\description{
Computes the default argument of various distributions. Adapted from fitdistrplus.
}
\examples{
x <- rnorm(100)
args <- start_arg_default(x, "norm")
args
}
